Improved planets

+++ Src/Physics/Gravity.cpp
@@ -12,9 +12,9 @@ Vector3 Gravity::Force(IBody *b1, IBody *b2)
 
     double r = (double)Vector3::Distance(pos1, pos2);
     double dx = pos1.x - pos2.x, dy = pos1.y - pos2.y, dz = pos1.z - pos2.z;
-    double f = (Constants::G * b1->GetMass() * b2->GetMass()) / (r * r);
+    double f = (Constants::G * (double)b1->GetMass() * (double)b2->GetMass()) / (r * r);
 
     f /= Constants::Scale;
 
-    return Vector3(f * dx / r, f * dy / r, f * dz / r);
+    return Vector3((float)(f * dx / r), (float)(f * dy / r), (float)(f * dz / r));
 }
